Update the use the unrelaxed_defect_structure
Add additional ordering matching step to ensure that the generated
unrelaxed defect structure mathches to that actually used in the
calculations. Then the `unrelaxed_defect_structure` can be used
as the `initial_defect_structure`.

pyCDT assuems `initial_defect_structure` does not undergo relaxation,
but for our workflow it may not be the case if multiple restarts
have taken place.


Former-commit-id: a0c9c1d642c1dae589c231cd7e179a4bcb50df64
Former-commit-id: e48cf1021e069b5bf574d419dec7a46c07754973