Guard import of `UserPotcarFunctional` behind an `if TYPE_CHECKING` block
This PR seeks to close https://github.com/SMTG-Bham/ShakeNBreak/issues/73, allowing `doped` to function with pymatgen>=5.31.2024.

In short, throughout Pymatgen, type hints are generally guarded behind an `if TYPE_CHECKING` block to ensure that compute time isn't wasted on loading imports during calculations. This is pretty standard practice.

Previously, `pymatgen.io.vasp.sets.UserPotcarFunctional` was not guarded by an `if TYPE_CHECKING` block and could be imported directly. However, in pymatgen==5.3.1.2024, it was guarded in the `if TYPE_CHECKING` block, meaning it can no longer be imported directly. This was the cause of the incompatability.

I have resolved the issue by guarding the import in doped now.